Add TCP support for writing logs (#9)
* Change severity levels to untyped

The severity levels are not int32, they are limited from 0-23.

* Refactor UDP code to make room for TCP feature

This is a breaking change. Clients will need to use gelf.NewUDPWriter
instead of gelf.NewWriter. The type returned will also change from
*gelf.Writer to *gelf.UDPWriter, or gelf.Writer for the interface.

* Add TCP transport support

We expose a NewTCPWriter function that returns a TCPWriter type
that implements the gelf.Writer interface.

* Avoid deadlock in the tcp reader

By using a channel instead of a pair of mutexes and a slice of bytes
attached to the reader.

This also allows to write more than one message over the same
connection.

* tcpreader: simplify the json decoding step